Move GeojsonMap to mapbox dir
